refactor: 로그 관련 수정
